+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * St-radio doc
 * Add table component example
 * Fix #212: Change type to header item
+* Add switch component example
 
 ## 1.3.0 (Abril 4, 2017)